+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## NEXT
+## [3.0.0]
 
 - **BREAKING CHANGES**:
   - Add currentIndex and previousIndex to the onSwipe callback
+  - Render the items through a builder function
 
 ## [2.1.0]